requirements/base.txt
pytz==2019.1  # https://github.com/stub42/pytz
python-slugify==3.0.2  # https://github.com/un33k/python-slugify
Pillow==6.1.0  # https://github.com/python-pillow/Pillow
argon2-cffi==19.1.0  # https://github.com/hynek/argon2_cffi
redis==3.2.1  # https://github.com/antirez/redis
celery==4.3.0  # pyup: < 5.0  # https://github.com/celery/celery
django-celery-beat==1.5.0  # https://github.com/celery/django-celery-beat
flower==0.9.3  # https://github.com/mher/flower

# Django
# ------------------------------------------------------------------------------
django==2.2.3  # pyup: < 3.0  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==3.2.0  # https://github.com/jazzband/django-model-utils
django-allauth==0.39.1  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.7.2  # https://github.com/django-crispy-forms/django-crispy-forms
django-redis==4.10.0  # https://github.com/niwinz/django-redis

# Django REST Framework
djangorestframework==3.9.0  # https://github.com/encode/django-rest-framework
django-rest-auth
coreapi==2.3.3  # https://github.com/core-api/python-client

# cvat requirements
click==6.7
django-appconf==1.0.2
django-cacheops==4.0.6
django-compressor==2.2
django-rq==2.0.0
EasyProcess==0.2.3
ffmpy==0.2.2
numpy==1.16.2
patool==1.12
pyunpack==0.1.2
rcssmin==1.0.6
requests==2.20.0
rjsmin==1.0.12
rq==1.0.0
scipy==1.2.1
sqlparse==0.2.4
django-sendfile==0.3.11
dj-pagination==2.4.0
python-logstash==0.4.6
django-revproxy==0.9.15
rules==2.0
GitPython==2.1.11
coreapi==2.3.3
django-filter==2.0.0
Markdown==3.0.1
Pygments==2.3.1
drf-yasg==1.16.0
Shapely==1.6.4.post2
furl==2.0.0
pascal_voc_writer==0.1.4
pdf2image==1.6.0
scikit-image>=0.14.0
tensorflow==1.14.0
pycocotools-fix
# python-ldap==3.0.0 # is causing trouble. need ldap2-dev package
